判题未通过导致无限判题的Bug已修复.

+++ judge-service/src/main/java/top/cloudli/judgeservice/task/FileCleaner.java
@@ -57,7 +57,9 @@ public class FileCleaner {
                 delete(f);
             }
         }
-        if (!file.delete())
-            log.info("{} 删除失败.", file.getName());
+        if (file.exists()) {
+            if (!file.delete())
+                log.info("{} 删除失败.", file.getName());
+        }
     }
 }
